common - shrink path UIDs from 64-bit to 32-bit to excape VS2013 64-bit atomics bug

+++ src-common/ast.h
@@ -42,7 +42,7 @@ class RendererAST;
 
 namespace AST {
     typedef std::vector<agg::trans_affine> SymmList;
-    typedef uint_fast64_t UIDdatatype;
+    typedef uint_fast32_t UIDdatatype;
 
     class ASTpath;
     class ASTreplacement;
